refactor scripts to packages
