+++ package.json
@@ -12,7 +12,7 @@
     "ng": "ng",
     "start": "ng serve --open",
     "build": "ng build --prod",
-    "scripts": "cd scripts && npm run start --",
+    "scripts": "cd packages/scripts && npm run start --",
     "server": "yarn workspace server",
     "test": "ng test",
     "lint": "ng lint",
