+++ packages/scripts/.gitignore
