+++ packages/scripts/config/credentials.json.enc
