+++ packages/scripts/config/public.key
