+++ packages/scripts/index.ts
