+++ packages/scripts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "parenting-app-scripts",
+  "name": "scripts",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
