+++ packages/scripts/src/decrypt-config.ts
