+++ packages/scripts/src/encrypt-config.ts
