+++ packages/scripts/src/gdrive-download/sample.env
