+++ packages/scripts/src/resizeAndroidAssets.ts
