+++ packages/scripts/src/sync-single.ts
