+++ packages/scripts/src/utils/cli-utils.ts
