+++ packages/scripts/src/utils/file-utils.ts
