+++ packages/scripts/src/utils/index.ts
@@ -2,4 +2,4 @@ export * from "./cli-utils";
 export * from "./file-utils";
 export * from "./string-utils";
 // TODO - make better way to share scripts and types between folders (/workspaces)
-export { extractDynamicFields } from "../../../src/app/shared/utils";
+export { extractDynamicFields } from "../../../../src/app/shared/utils";
