+++ packages/scripts/src/utils/string-utils.ts
