+++ packages/scripts/src/version.ts
