+++ packages/scripts/tsconfig.json
