+++ packages/scripts/types/index.ts
@@ -1,0 +1,2 @@
+// Provide access to FlowTypes used in the app
+export { FlowTypes, RapidProFlowExport } from "../../../src/app/shared/model/flowTypes";
