--- scripts/types/index.ts
@@ -1,2 +1,0 @@
-// Provide access to FlowTypes used in the app
-export { FlowTypes, RapidProFlowExport } from "../../src/app/shared/model/flowTypes";
