+++ src/data/habit_list/index.ts
@@ -1,3 +1,3 @@
-import { FlowTypes } from "src/app/shared/model/flowTypes";
+import { FlowTypes } from "src/app/shared/model/FlowTypes";
 import habit_list_0 from "./habit_list";
 export const habit_list:FlowTypes.Habit_list[] = [].concat(habit_list_0)
